Refactor tokenizer
